+++ src/lib/Tcp/TcpCommon.ts
@@ -34,3 +34,10 @@ export const DEFAULT_HOSTNAME: string = 'localhost';
  * @type uint32
  */
 export const DEFAULT_BACKLOG: number = 1023;
+
+/**
+ * The default timeout for connecting to the server, in milliseconds.
+ *
+ * @type uint32
+ */
+export const DEFAULT_CONNECT_TIMEOUT: number = 30_000;